clipmind_ai/requirements.txt
PySide6>=6.6.1
mss>=9.0.1
paddleocr>=2.8.0
paddlepaddle>=2.6.2
Pillow>=10.1.0
requests>=2.31.0
httpx>=0.26.0
keyboard>=0.13.5
pywin32>=306
beautifulsoup4>=4.12.2
trafilatura>=1.6.3
pydantic>=2.5.3
loguru>=0.7.2
sqlalchemy>=2.0.25
pyinstaller>=6.3.0
pyperclip>=1.8.2
